- Add a UiSate to the ViewModel of the HomeScreen. - Refactoring the logic from the screen to the ViewModel and the UiState
