+++ app/src/main/java/com/alura/aluvery/ui/activities/MainActivity.kt
@@ -4,6 +4,7 @@ import android.content.Intent
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
+import androidx.activity.viewModels
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material.FloatingActionButton
@@ -15,15 +16,13 @@ import androidx.compose.material.icons.filled.Add
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.alura.aluvery.dao.ProductDao
 import com.alura.aluvery.sampledata.sampleSections
 import com.alura.aluvery.ui.screens.HomeScreen
-import com.alura.aluvery.ui.screens.HomeScreenUiState
+import com.alura.aluvery.ui.states.HomeScreenUiState
 import com.alura.aluvery.ui.theme.AluveryTheme
+import com.alura.aluvery.ui.viewmodels.HomeScreenViewModel
 
 class MainActivity : ComponentActivity() {
-
-    private val dao = ProductDao()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -32,8 +31,8 @@ class MainActivity : ComponentActivity() {
                 startActivity(Intent(this, ProductFormActivity::class.java))
             },
                 content = {
-                    val products = dao.products()
-                    HomeScreen(products = products)
+                    val viewModel by viewModels<HomeScreenViewModel>()
+                    HomeScreen(viewModel)
                 }
             )
         }
