+++ app/src/main/java/com/alura/aluvery/ui/states/HomeScreenUiState.kt
@@ -1,0 +1,12 @@
+package com.alura.aluvery.ui.states
+
+import com.alura.aluvery.model.Product
+
+data class HomeScreenUiState(
+    val sections: Map<String, List<Product>> = emptyMap(),
+    val searchedProducts: List<Product> = emptyList(),
+    val searchText: String = "",
+    val onSearchChange: (String) -> Unit = {}
+) {
+    fun isShownSections() = searchText.isBlank()
+}
